+++ raytracing/ray.cpp
@@ -4,6 +4,6 @@ ray::ray() : origin{ 0,0,0 }, direction{ 0,0,0 } {}
 
 ray::ray(const vector3& o, const vector3& d) : origin(o), direction(d) {}
 
-vector3 ray::at(float t) const {
+vector3 ray::at(double t) const {
 	return origin + direction * t;
 }
