+++ raytracing/sphere.h
@@ -6,12 +6,12 @@ class sphere :
     public hittable
 {
 public:
-    sphere(vector3 cen, float r) : center(cen), radius(r) {};
+    sphere(vector3 cen, double r) : center(cen), radius(r) {};
 
     virtual bool hit(
-        const ray& r, float t_min, float t_max, hit_record& rec) const override;
+        const ray& r, double t_min, double t_max, hit_record& rec) const override;
 
 public:
     vector3 center;
-    float radius;
+    double radius;
 };
